Start to add svg shadows
No visual change yet. Currently just rendering a clear svg path
