+++ lib/chukinas_web/templates/dreadnought/_sprite.html.leex
@@ -1,18 +1,25 @@
 <%# Sprite Template %>
 <div 
   class="relative overflow-hidden"
-  style="<%= ChukinasWeb.Shared.width_height_from_size(@rect) %>"
->
-<div
+  style="<%= ChukinasWeb.Shared.width_height_from_size(@rect) %>">
+  <div
     class="absolute"
     style="
       <%= ChukinasWeb.Shared.width_height_from_size(@image_size) %>
       <%= ChukinasWeb.Shared.left_top_from_position(@transform) %>
-    "
-  >
-  <img 
-    style="clip-path: path('<%= @image_clip_path %>');"
-    src="<%= Routes.static_path @socket, @image_file_path %>"
-  >
+    ">
+    <img 
+      style="clip-path: path('<%= @image_clip_path %>');"
+      src="<%= Routes.static_path @socket, @image_file_path %>">
+  </div>
+  <div
+    class="absolute"
+    style="
+      <%= ChukinasWeb.Shared.width_height_from_size(@image_size) %>
+      <%= ChukinasWeb.Shared.left_top_from_position(@transform) %>
+    ">
+    <svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" viewBox="<%= @image_viewbox %>">
+      <path d="<%= @image_clip_path %>" fill="#FF555500"/>
+    </svg>
   </div>
 </div>
