+++ lib/chukinas_web/views/shared_components.ex
@@ -20,21 +20,3 @@ defmodule ChukinasWeb.SharedComponents do
   end
 
 end
-
-defmodule ChukinasWeb.Shared do
-
-  alias ChukinasWeb.DreadnoughtPlayView
-
-  def top_left_width_height_from_rect(rect) do
-    DreadnoughtPlayView.render_rect_as_style_attrs(rect)
-  end
-
-  def left_top_from_position(position) do
-    DreadnoughtPlayView.left_top_from_position position
-  end
-
-  def width_height_from_size(size) do
-    DreadnoughtPlayView.width_height_from_size size
-  end
-
-end
